Corrected job vacancies creation / editing due to changes in category structure.
I also recreated the migration file.

+++ database/migrations/2020_10_25_052812_create_apply_details_table.php
@@ -15,7 +15,7 @@ class CreateApplyDetailsTable extends Migration
     {
         Schema::create('apply_details', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('apply_id')->index();
+            $table->unsignedBigInteger('apply_id')->index();
             $table->foreign('apply_id')->references('id')->on('applies')->onDelete('cascade');
             $table->string('last_name');
             $table->string('first_name');
